requirements.txt
aiofile==3.8.8
anyio==4.4.0
asgiref==3.7.2
attrs==24.1.0
autobahn==24.4.2
Automat==22.10.0
autopep8==2.3.1
caio==0.9.17
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
colorlog==6.8.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
dill==0.3.8
Django==5.0
djangorestframework==3.15.2
fastjsonschema==2.20.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jsonpatch==1.33
jsonpointer==3.0.0
lxml==5.2.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
orjson==3.10.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pyOpenSSL==24.2.1
reactpy==1.0.2
reactpy-router==0.1.1
reactpy_django==4.0.0
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.2.2
zope.interface==6.4.post2
